multi-service responses implemented:
The problem is can potentially have many services, and these services all want to return something.
Those "somethings" are one of two things:
1. JSON formatted results
2. A full HTTP response (such as pre-gzipped HTML, etc)

We attempted to solve this generically by sending back multipart formdata http responses, but it
is not possible to send these back generically enough (some might want to be gzipped while others
might not want to be, for example). Our compromise is the following system:

1. All JSON responses are gathered together into a single array (or a single dictionary if there is only one).
   This combined JSON is attached to an http header "Service-Response"
2. If none of the services include an http response, we make a simple 0 content success and send it
3. If there is exactly one http response, we attach the headers from step 1 and send it
4. If there are more then one http response, we send internal service error and attach header from step 1 and send it